+++ service.go
@@ -7,9 +7,7 @@ import (
 	"io/ioutil"
 	"net/http"
 	"net/url"
-	"strconv"
 	"strings"
-	"time"
 )
 
 var ClientId int
@@ -106,34 +104,6 @@ func (client *Client) runRequest(req *http.Request) ([]byte, error) {
 	updateRateLimits(resp)
 
 	return checkResponseForErrors(resp)
-}
-
-func updateRateLimits(resp *http.Response) error {
-	var err error
-
-	if resp.Header.Get("X-Ratelimit-Limit") == "" || resp.Header.Get("X-Ratelimit-Usage") == "" {
-		return errors.New("ratelimit headers not found")
-	}
-
-	s := strings.Split(resp.Header.Get("X-Ratelimit-Limit"), ",")
-	if RateLimitLast.LimitShort, err = strconv.Atoi(s[0]); err != nil {
-		return err
-	}
-	if RateLimitLast.LimitLong, err = strconv.Atoi(s[1]); err != nil {
-		return err
-	}
-
-	s = strings.Split(resp.Header.Get("X-Ratelimit-Usage"), ",")
-	if RateLimitLast.UsageShort, err = strconv.Atoi(s[0]); err != nil {
-		return err
-	}
-	if RateLimitLast.UsageLong, err = strconv.Atoi(s[1]); err != nil {
-		return err
-	}
-
-	RateLimitLast.LastRequestTime = time.Now()
-
-	return nil
 }
 
 func checkResponseForErrors(resp *http.Response) ([]byte, error) {
